Rework player sprite handling
All player sprites (overworld, front and back pics etc) have been reworked so they're in a table (gPlayerSpriteTable) and is done in a way where all a person would need to do is change their version to use the respective player sprites. Link player has also been changed so it accesses the table.

Also organised custom overworld palettes a bit and removed a pointless field in the object event graphics info struct

+++ include/pokemon.h
@@ -527,7 +527,7 @@ const u8 *GetTrainerNameFromId(u16 trainerId);
 struct MonSpritesGfxManager *CreateMonSpritesGfxManager(void);
 void DestroyMonSpritesGfxManager(void);
 u8 *MonSpritesGfxManager_GetSpritePtr(void);
-u8 PlayerGenderToBackTrainerPicId(u8 playerGender);
+u16 PlayerGenderToBackTrainerPicId(u8 playerGender);
 const u8 *GetSpeciesNameOfLanguage(u16 species, u8 language);
 bool32 IsMonNotNicknamed(u8 *nickname, u16 species, u8 language);
